# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Agbalumo API
  description: |
    REST API for the Agbalumo directory platform - a web application for the West African diaspora community.
    
    ## Authentication
    - Session-based authentication using cookies
    - Google OAuth 2.0 for social login
    - CSRF protection required for state-changing operations
    
    ## Rate Limiting
    - Default: 60 requests per minute
    - Admin login: 5 requests per minute
  version: 1.0.0
  contact:
    name: Agbalumo Team
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://agbalumo.com
    description: Production server
...
paths:
...
  /admin:
    get:
      summary: Admin dashboard
      description: Returns admin dashboard with statistics
      tags:
        - Admin
      security:
        - CookieAuth: []
      responses:
        '200':
          description: Admin dashboard HTML
        '302':
          description: Redirect to login
  /admin/login:
    get:
      summary: Admin login form
      tags:
        - Admin
      responses:
        '200':
          description: Login form HTML

    post:
      summary: Admin login action
      tags:
        - Admin
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - code
              properties:
                code:
                  type: string
      responses:
        '200':
          description: Login result
        '400':
          description: Invalid code
...
  /admin/listings:
    get:
      summary: List all listings (admin)
      description: Returns all listings with moderation controls
      tags:
        - Admin
      security:
        - CookieAuth: []
      parameters:
        - name: category
          in: query
          description: Filter by category
          schema:
            type: string
        - name: sort
          in: query
          description: Sort field
          schema:
            type: string
        - name: order
          in: query
          description: Sort order
          schema:
            type: string
            enum: [ASC, DESC]
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
      responses:
        '200':
          description: Listings list HTML
...
  /admin/listings/delete-confirm:
    get:
      summary: Delete confirmation
      description: Returns delete confirmation modal
      tags:
        - Admin
      security:
        - CookieAuth: []
      parameters:
        - name: id
          in: query
          description: ID(s) to delete
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Confirmation modal HTML
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: session
...